refactor(api): move assessment-result routes and controller
