+++ api/src/shared/routes.js
@@ -1,5 +1,6 @@
 import * as assessmentsRoutes from './application/assessments/index.js';
+import * as assessmentResultsRoutes from './application/assessment-results/index.js';
 
-const sharedRoutes = [assessmentsRoutes];
+const sharedRoutes = [assessmentsRoutes, assessmentResultsRoutes];
 
 export { sharedRoutes };
